Login do cliente

+++ app/views/cliente.blade.php
@@ -31,13 +31,15 @@
             <tr>
               <td scope="row">{{ $cliente->getNome() }}</td>
               <td>
-                <a href="/cliente/editar/{{ $cliente->getId() }}" class="btn btn-xs btn-success">Editar</a>
-                @if($cliente->getAtivo())
-                    <a href="/cliente/desativar/{{ $cliente->getId() }}" class="btn btn-xs btn-warning">Desativar</a>
-                @else 
-                    <a href="/cliente/ativar/{{ $cliente->getId() }}" class="btn btn-xs btn-primary">Ativar</a>
+                <a href="/cliente/view/{{ $cliente->getId() }}" class="btn btn-xs btn-success">Visualizar</a>
+                @if(Auth::user()->tipo == 'admin')
+                    @if($cliente->getAtivo())
+                        <a href="/cliente/desativar/{{ $cliente->getId() }}" class="btn btn-xs btn-warning">Desativar</a>
+                    @else 
+                        <a href="/cliente/ativar/{{ $cliente->getId() }}" class="btn btn-xs btn-primary">Ativar</a>
+                    @endif
+                    <a href="/cliente/excluir/{{ $cliente->getId() }}" class="btn btn-xs btn-danger">Excluir</a>
                 @endif
-                <a href="/cliente/excluir/{{ $cliente->getId() }}" class="btn btn-xs btn-danger">Excluir</a>
               </td>
             </tr>
         @endforeach
